requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
elastic-transport==8.17.1
elasticsearch==8.18.0
elasticsearch-dsl==8.18.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.116.0
ffmpeg-python==0.2.0
fonttools==4.57.0
future==1.0.0
greenlet==3.2.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
json_repair==0.50.1
kiwisolver==1.4.8
matplotlib==3.10.1
mysql-connector-python==9.4.0
neo4j==5.28.2
numpy==2.2.4
openai==1.65.2
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
patsy==1.0.1
pefile==2023.2.7
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
starlette==0.46.2
statsmodels==0.14.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
